+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-08-13 19:02:32
-// Commit: test(auth): add unit tests for auth business logic
+// Updated at: 2026-08-14 12:32:22
+// Commit: feat(auth): add bi-directional event bus handler for auth
 
 class authComponentState {
-  final String commitRef = '3590';
-  final String lastMessage = 'test(auth): add unit tests for auth business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-13T19:02:32+05:00');
+  final String commitRef = '3606';
+  final String lastMessage = 'feat(auth): add bi-directional event bus handler for auth';
+  final DateTime timestamp = DateTime.parse('2026-08-14T12:32:22+05:00');
 }